# src/main/resources/swagger.yaml
---
swagger: 2.0
info:
  description: "This is a sample Petstore server.  You can find \nout more about Swagger at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  version: 1.0.0
  title: Swagger Petstore
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /ignaciogarciav/petstore/1.0.0
tags:
- name: pet
  description: Everything about your Pets
  externalDocs:
    description: Find out more
    url: http://swagger.io
- name: store
  description: Access to Petstore orders
- name: user
  description: Operations about user
  externalDocs:
    description: Find out more about our store
    url: http://swagger.io
schemes:
- https
- http
paths:
  /pet:
    post:
      tags:
      - pet
      summary: Add a new pet to the store
      operationId: addPet
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: Pet object that needs to be added to the store
        required: true
        schema:
          $ref: '#/definitions/Pet'
      responses:
        "201":
          description: Pet successfully added
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: invalid fields
          schema:
            $ref: '#/definitions/ResponseApi'
        "405":
          description: Invalid input
  /pet/findByStatus:
    get:
      tags:
      - pet
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      produces:
      - application/json
      - application/xml
      parameters:
      - name: status
        in: query
        description: Status values that need to be considered for filter
        required: true
        type: array
        items:
          type: string
          default: available
          enum:
          - available
          - pending
          - sold
        collectionFormat: multi
      responses:
        "200":
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Pet'
        "400":
          description: Invalid status value
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: No pets with desired status
          schema:
            $ref: '#/definitions/ResponseApi'
  /pet/findByTags:
    get:
      tags:
      - pet
      summary: Finds Pets by tags
      description: Muliple tags can be provided with comma separated strings. Use\ \ tag1, tag2, tag3 for testing.
      operationId: findPetsByTags
      produces:
      - application/json
      - application/xml
      parameters:
      - name: tags
        in: query
        description: Tags to filter by
        required: true
        type: array
        items:
          type: string
        collectionFormat: multi
      responses:
        "200":
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Pet'
        "400":
          description: Invalid tag value
      deprecated: true
  /pet/{petId}:
    get:
      tags:
      - pet
      summary: Find pet by ID
      description: Returns a single pet
      operationId: getPetById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: petId
        in: path
        description: ID of pet to return
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Pet'
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Pet not found
          schema:
            $ref: '#/definitions/ResponseApi'
    post:
      tags:
      - pet
      summary: Updates a pet in the store with form data
      operationId: updatePetWithForm
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      - application/xml
      parameters:
      - name: petId
        in: path
        description: ID of pet that needs to be updated
        required: true
        type: integer
        format: int64
      - name: name
        in: formData
        description: Updated name of the pet
        required: false
        type: string
      - name: status
        in: formData
        description: Updated status of the pet
        required: false
        type: string
      responses:
        "200":
          description: Successful update
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid fields
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Pet not found
          schema:
            $ref: '#/definitions/ResponseApi'
        "405":
          description: Invalid input
    put:
      tags:
      - pet
      summary: Update an existing pet
      operationId: updatePet
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - name: petId
        in: path
        description: ID of pet that needs to be updated
        required: true
        type: integer
        format: int64
      - in: body
        name: body
        description: Pet object that needs to be added to the store
        required: true
        schema:
          $ref: '#/definitions/Pet'
      responses:
        "200":
          description: Successful update
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Pet not found
          schema:
            $ref: '#/definitions/ResponseApi'
        "405":
          description: Validation exception
          schema:
            $ref: '#/definitions/ResponseApi'
    delete:
      tags:
      - pet
      summary: Deletes a pet
      operationId: deletePet
      produces:
      - application/json
      - application/xml
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: petId
        in: path
        description: Pet id to delete
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successfully deleted
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Pet not found
          schema:
            $ref: '#/definitions/ResponseApi'
  /pet/{petId}/uploadImage:
    post:
      tags:
      - pet
      summary: uploads an image
      operationId: uploadFile
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - name: petId
        in: path
        description: ID of pet to update
        required: true
        type: integer
        format: int64
      - name: additionalMetadata
        in: formData
        description: Additional data to pass to server
        required: false
        type: string
      - name: file
        in: formData
        description: file to upload
        required: false
        type: file
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid field
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Pet not found
          schema:
            $ref: '#/definitions/ResponseApi'
  /store/inventory:
    get:
      tags:
      - store
      summary: Returns pet inventories by status
      description: Returns a map of status codes to quantities
      operationId: getInventory
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: successful operation
          schema:
            type: object
            additionalProperties:
              type: integer
              format: int32
  /store/order:
    post:
      tags:
      - store
      summary: Place an order for a pet
      operationId: placeOrder
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: order placed for purchasing the pet
        required: true
        schema:
          $ref: '#/definitions/Order'
      responses:
        "201":
          description: successful operation
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid Order
          schema:
            $ref: '#/definitions/ResponseApi'
  /store/order/{orderId}:
    get:
      tags:
      - store
      summary: Find purchase order by ID
      description: For valid response try integer IDs with value >= 1 and <= 10.\ \ Other values will generated exceptions
      operationId: getOrderById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: orderId
        in: path
        description: ID of pet that needs to be fetched
        required: true
        type: integer
        maximum: 10
        minimum: 1
        format: int64
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/ResponseApi'
    delete:
      tags:
      - store
      summary: Delete purchase order by ID
      description: For valid response try integer IDs with positive integer value.\ \ Negative or non-integer values will generate API errors
      operationId: deleteOrder
      produces:
      - application/json
      - application/xml
      parameters:
      - name: orderId
        in: path
        description: ID of the order that needs to be deleted
        required: true
        type: integer
        minimum: 1
        format: int64
      responses:
        "200":
          description: order sucessfully deleted
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/ResponseApi'
  /user:
    post:
      tags:
      - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: Created user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        "201":
          description: User successfully created
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid fields
          schema:
            $ref: '#/definitions/ResponseApi'
  /user/createWithArray:
    post:
      tags:
      - user
      summary: Creates list of users with given input array
      operationId: createUsersWithArrayInput
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: List of user object
        required: true
        schema:
          type: array
          items:
            $ref: '#/definitions/User'
      responses:
        "201":
          description: User successfully created
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid fields
          schema:
            $ref: '#/definitions/ResponseApi'
  /user/createWithList:
    post:
      tags:
      - user
      summary: Creates list of users with given input array
      operationId: createUsersWithListInput
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: List of user object
        required: true
        schema:
          type: array
          items:
            $ref: '#/definitions/User'
      responses:
        "201":
          description: User successfully created
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid fields
          schema:
            $ref: '#/definitions/ResponseApi'
  /user/login:
    get:
      tags:
      - user
      summary: Logs user into the system
      operationId: loginUser
      produces:
      - application/json
      - application/xml
      parameters:
      - name: username
        in: query
        description: The user name for login
        required: true
        type: string
      - name: password
        in: query
        description: The password for login in clear text
        required: true
        type: string
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid username/password supplied
          schema:
            $ref: '#/definitions/ResponseApi'
  /user/logout:
    get:
      tags:
      - user
      summary: Logs out current logged in user session
      operationId: logoutUser
      produces:
      - application/json
      - application/xml
      parameters: []
      responses:
        "200":
          description: Successful logout
          schema:
            $ref: '#/definitions/ResponseApi'
  /user/{username}:
    get:
      tags:
      - user
      summary: Get user by user name
      operationId: getUserByName
      produces:
      - application/json
      - application/xml
      parameters:
      - name: username
        in: path
        description: The name that needs to be fetched. Use user1 for testing.
        required: true
        type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Invalid username supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ResponseApi'
    put:
      tags:
      - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      produces:
      - application/json
      - application/xml
      parameters:
      - name: username
        in: path
        description: name that need to be updated
        required: true
        type: string
      - in: body
        name: body
        description: Updated user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        "200":
          description: User successfully updated
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid user supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ResponseApi'
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      produces:
      - application/json
      - application/xml
      parameters:
      - name: username
        in: path
        description: The name that needs to be deleted
        required: true
        type: string
      responses:
        "200":
          description: User successfully deleted
          schema:
            $ref: '#/definitions/ResponseApi'
        "400":
          description: Invalid username supplied
          schema:
            $ref: '#/definitions/ResponseApi'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ResponseApi'
definitions:
  ResponseApi:
    type: object
    properties:
      timestamp:
        type: string
        format: date-time
      status:
        type: string
        description: HttpStatus
        enum:
        - 200 OK
        - 201 CREATED
        - 400 BAD_REQUEST
        - 404 NOT_FOUND
        - 500 INTERNAL_SERVER_ERROR
      message:
        type: string
  Order:
    type: object
    properties:
      id:
        type: integer
        format: int64
      petId:
        type: integer
        format: int64
      quantity:
        type: integer
        format: int32
      shipDate:
        type: string
        format: date-time
      status:
        type: string
        description: Order Status
        enum:
        - placed
        - approved
        - delivered
      complete:
        type: boolean
        default: false
    xml:
      name: Order
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Category
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
        format: email
      password:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int32
        description: User Status
    xml:
      name: User
  Tag:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Tag
  Pet:
    type: object
    required:
    - name
    - photoUrls
    properties:
      id:
        type: integer
        format: int64
      category:
        $ref: '#/definitions/Category'
      name:
        type: string
        example: doggie
      photoUrls:
        type: array
        xml:
          name: photoUrl
          wrapped: true
        items:
          type: string
      tags:
        type: array
        xml:
          name: tag
          wrapped: true
        items:
          $ref: '#/definitions/Tag'
      status:
        type: string
        description: pet status in the store
        enum:
        - available
        - pending
        - sold
    xml:
      name: Pet
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io